rename `applet` module to `applets`
