+++ ctru-rs/src/applets/mod.rs
