+++ ctru-rs/src/lib.rs
@@ -11,7 +11,7 @@ extern crate widestring;
 
 extern crate ctru_sys as libctru;
 
-pub mod applet;
+pub mod applets;
 pub mod console;
 pub mod error;
 pub mod srv;
